# backend/services/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AlertConfig:
    properties:
      cooldown_minutes:
        type: integer
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      notification_channels:
        items:
          type: integer
        type: array
      project_id:
        type: string
      spike_threshold:
        type: integer
      spike_window_minutes:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.BrandCatalog:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      source_url:
        type: string
    type: object
  domain.Event:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      metadata:
        items:
          type: integer
        type: array
      project_id:
        type: string
      source_id:
        type: string
      type:
        type: string
    type: object
...
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.registerRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  handler.upsertAlertConfigRequest:
    properties:
      cooldown_minutes:
        minimum: 1
        type: integer
      enabled:
        type: boolean
      notification_channels:
        items:
          type: integer
        type: array
      spike_threshold:
        minimum: 1
        type: integer
      spike_window_minutes:
        minimum: 1
        type: integer
    required:
    - cooldown_minutes
    - enabled
    - spike_threshold
    - spike_window_minutes
    type: object
  httputil.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  repo.AlertRow:
    properties:
      created_at:
        type: string
      id:
        type: string
      mention_count:
        type: integer
      project_id:
        type: string
      sent_at:
        type: string
      sentiment_summary:
        items:
          type: integer
        type: array
      status:
        type: string
      threshold:
        type: integer
      type:
        type: string
    type: object
...
  service.AuthTokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Brand reputation monitoring system API
  title: BrandRadar API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AuthTokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AuthTokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.AuthTokens'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /brand-catalog:
    get:
      parameters:
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List brand catalog
      tags:
      - brand-catalog
  /brand-catalog/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BrandCatalog'
            type: array
      summary: Search brand catalog
      tags:
      - brand-catalog
...
  /projects/{projectID}/alert-config:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AlertConfig'
      security:
      - BearerAuth: []
      summary: Get alert configuration
      tags:
      - alert-config
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Alert config data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.upsertAlertConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AlertConfig'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create or update alert configuration
      tags:
      - alert-config
  /projects/{projectID}/alerts:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.AlertRow'
            type: array
      security:
      - BearerAuth: []
      summary: List project alerts
      tags:
      - alerts
...
  /projects/{projectID}/events:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
      security:
      - BearerAuth: []
      summary: List project events
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"